organize growing number of JpaRepositoruies
