+++ src/main/java/org/openhds/repository/concrete/LocationHierarchyLevelRepository.java
@@ -1,10 +1,11 @@
-package org.openhds.repository;
+package org.openhds.repository.concrete;
 
 /**
  * Created by Ben on 5/4/15.
  */
 
 import org.openhds.domain.model.LocationHierarchyLevel;
+import org.openhds.repository.contract.AuditableRepository;
 
 import java.util.Optional;
