+++ src/main/java/org/openhds/repository/concrete/PrivilegeRepository.java
@@ -1,4 +1,4 @@
-package org.openhds.repository;
+package org.openhds.repository.concrete;
 
 /**
  * Created by Ben on 5/4/15.
