+++ src/main/java/org/openhds/repository/concrete/UserRepository.java
@@ -1,9 +1,10 @@
-package org.openhds.repository;
+package org.openhds.repository.concrete;
 
 /**
  * Created by Ben on 5/4/15.
  */
 
+import org.openhds.repository.contract.UuidIdentifiableRepository;
 import org.openhds.security.model.User;
 
 import java.util.List;
