+++ src/main/java/org/openhds/repository/contract/AuditableExtIdRepository.java
@@ -1,4 +1,4 @@
-package org.openhds.repository;
+package org.openhds.repository.contract;
 
 import org.openhds.domain.contract.AuditableExtIdEntity;
 import org.springframework.data.repository.NoRepositoryBean;
