+++ src/main/java/org/openhds/service/contract/AbstractAuditableExtIdService.java
@@ -1,7 +1,7 @@
 package org.openhds.service.contract;
 
 import org.openhds.domain.contract.AuditableExtIdEntity;
-import org.openhds.repository.AuditableExtIdRepository;
+import org.openhds.repository.contract.AuditableExtIdRepository;
 
 /**
  * Created by wolfe on 6/11/15.
