+++ src/main/java/org/openhds/service/contract/AbstractUuidService.java
@@ -1,11 +1,10 @@
 package org.openhds.service.contract;
 
 import org.openhds.domain.contract.UuidIdentifiable;
-import org.openhds.repository.UuidIdentifiableRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.openhds.repository.contract.UuidIdentifiableRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.List;
 
 /**
