+++ src/test/java/org/openhds/resource/controller/LocationRestControllerTest.java
@@ -1,9 +1,9 @@
 package org.openhds.resource.controller;
 
 import org.openhds.domain.model.Location;
-import org.openhds.repository.FieldWorkerRepository;
-import org.openhds.repository.LocationHierarchyRepository;
-import org.openhds.repository.LocationRepository;
+import org.openhds.repository.concrete.FieldWorkerRepository;
+import org.openhds.repository.concrete.LocationHierarchyRepository;
+import org.openhds.repository.concrete.LocationRepository;
 import org.openhds.resource.contract.AuditableExtIdRestControllerTest;
 import org.openhds.resource.registration.LocationRegistration;
 import org.openhds.resource.registration.Registration;
