+++ src/test/java/org/openhds/resource/controller/UserRestControllerTest.java
@@ -1,6 +1,6 @@
 package org.openhds.resource.controller;
 
-import org.openhds.repository.UserRepository;
+import org.openhds.repository.concrete.UserRepository;
 import org.openhds.resource.contract.UuidIdentifiableRestControllerTest;
 import org.openhds.resource.registration.Registration;
 import org.openhds.resource.registration.UserRegistration;
